Quick POC
This is a POC of usage of shuffle webhook as entry for a responder for thehive.

I am going to add extra layer to it later but this may help others to create some quick responders. I see for example an use in things like EDR not yet supported by cortex but included in Shuffle.